collapsible functionality attributed to all cards
